Create SpanExporter with CDI. PR review
